requirements.txt
streamlit
langchain
langchain-core
langchain-openai
langchain-community
langchain-huggingface
langchain-pinecone
pinecone-client
python-dotenv
PyPDF2 
langgraph